backend/requirements.txt
alembic==1.14.1
aiohttp==3.10.10
fastapi==0.115.6
google-auth>=2.35.0,<3.0.0
google-genai>=0.8.0,<2.0.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
prometheus-client==0.21.1
PyJWT==2.10.1
pydantic==2.10.4
python-dotenv==1.0.1
redis==5.2.1
SQLAlchemy==2.0.37
uvicorn==0.32.1
